feat: change modal in profile component
