--- src/components/Modal/Modal.module.css
@@ -1,53 +1,0 @@
-.modal-close-btn {
-  float: right;
-  border: 2px solid #000000;
-  padding: 1px;
-  cursor: pointer;
-}
-
-.modal-close-btn:hover {
-  cursor: pointer;
-}
-
-.modal {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  width: 100%;
-  position: fixed;
-  top: 0;
-  left: 0;
-  background-color: rgba(0, 0, 0, 0.3);
-  opacity: 0;
-  pointer-events: none;
-  transition: 0.3s;
-
-}
-
-.active {
-  opacity: 1;
-  pointer-events: all;
-}
-
-.content {
-  background-color: #ffffff;
-  font-weight: 600;
-  font-size: 1.5rem;
-  padding: 20px;
-  border-radius: 10px;
-}
-.confirm__delete{
-  background-color: red;
-}
-.close{
-  cursor: pointer;
-  font-size: 1.5rem;
-  text-align: end;
-}
-.message__wrapper{
-  margin-top: 20px;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
