+++ src/components/ModalConfirm/ModalConfirm.tsx
@@ -1,12 +1,15 @@
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
-import { deleteBoard, deleteColumn } from '../../utils/fetch';
+import { deleteBoard, deleteColumn, deleteUser } from '../../utils/fetch';
 import { PropsModalConfirm } from '../../utils/types';
+import { clearLocalStorage } from '../../utils/utils';
+import { useAuth } from '../hook/useAuth';
 import styles from './ModalConfirm.module.css';
 
 export function ModalConfirm(props: PropsModalConfirm) {
   const { t } = useTranslation();
   const navigate = useNavigate();
+  const [state, dispatch] = useAuth();
 
   const handleDeleteColumn = async () => {
     try {
@@ -29,6 +32,28 @@ export function ModalConfirm(props: PropsModalConfirm) {
     }
   };
 
+  async function handleDeleteAccount() {
+    try {
+      const response = await deleteUser(state.id, state.token);
+      if (response.status > 399) {
+        throw new Error(`Something went wrong... Error code: ${response.status}`);
+      }
+      clearLocalStorage();
+      dispatch({
+        type: 'user',
+        data: {
+          username: null,
+          login: null,
+          token: null,
+          id: null,
+        },
+      });
+      navigate('/');
+    } catch (error) {
+      props.setError('Error');
+    }
+  }
+
   return (
     <div
       className={props.active ? `${styles.modal} ${styles.active}` : `${styles.modal}`}
@@ -46,10 +71,18 @@ export function ModalConfirm(props: PropsModalConfirm) {
             </button>
           </div>
         )}
-        {props.columnId !== 'deleteBoard' && (
+        {props.columnId !== 'deleteBoard' && props.columnId !== 'deleteUser' && (
           <div className={styles.message__wrapper}>
             {t('Delete column')}?
             <button className={styles.confirm__delete} onClick={handleDeleteColumn}>
+              {t('Confirm')}
+            </button>
+          </div>
+        )}
+        {props.columnId === 'deleteUser' && (
+          <div className={styles.message__wrapper}>
+            {t('Delete current user')}?
+            <button className={styles.confirm__delete} onClick={handleDeleteAccount}>
               {t('Confirm')}
             </button>
           </div>
